+++ Models/OrderPart.cs
@@ -63,8 +63,8 @@ namespace Nwazet.Commerce.Models {
         public void Build(
             ICharge charge,
             IEnumerable<CheckoutItem> items,
-            double subTotal,
-            double total,
+            decimal subTotal,
+            decimal total,
             TaxAmount taxes,
             ShippingOption shippingOption,
             Address shippingAddress,
@@ -73,13 +73,13 @@ namespace Nwazet.Commerce.Models {
             string customerPhone,
             string specialInstructions, 
             string currencyCode,
-            double amountPaid = default(double),
+            decimal amountPaid = default(decimal),
             string purchaseOrder = "") {
 
             _contentDocument = new XElement(ContentName)
                 .Attr(SubtotalName, subTotal)
                 .Attr(TotalName, total)
-                .Attr(AmountName, amountPaid == default(double) ? total : amountPaid)
+                .Attr(AmountName, amountPaid == default(decimal) ? total : amountPaid)
                 .Attr(PurchaseOrderName, purchaseOrder)
                 .Attr(CurrencyCodeName, currencyCode)
                 .AddEl(new XElement(ChargeName).With(charge)
@@ -185,9 +185,9 @@ namespace Nwazet.Commerce.Models {
             }
         }
 
-        public double SubTotal {
+        public decimal SubTotal {
             get {
-                return (double) ContentDocument.Attribute(SubtotalName);
+                return (decimal) ContentDocument.Attribute(SubtotalName);
             }
         }
 
@@ -204,17 +204,17 @@ namespace Nwazet.Commerce.Models {
             }
         }
 
-        public double Total {
+        public decimal Total {
             get {
-                return (double) ContentDocument.Attribute(TotalName);
+                return (decimal) ContentDocument.Attribute(TotalName);
             }
         }
 
-        public double AmountPaid {
+        public decimal AmountPaid {
             get {
                 var attr = ContentDocument.Attribute(AmountName);
                 if (attr == null) return Total;
-                return (double) attr;
+                return (decimal) attr;
             }
             set {
                 ContentDocument.SetAttributeValue(AmountName, value);
